requirement.txt
numpy
RPI.GPIO
paho-mqtt
aioconsole